+++ src/Stack/Runners.hs
@@ -185,7 +185,7 @@ withEnvConfigExt
     -> BuildOptsCLI
     -> Maybe (RIO BuildConfig ())
     -- ^ Action to perform before the build.  This will be run on the host
-    -- OS even if Docker is enabled for builds.  The build config is not
+    -- OS even if Docker is enabled for builds.  The env config is not
     -- available in this action, since that would require build tools to be
     -- installed on the host OS.
     -> (Maybe FileLock -> RIO EnvConfig a)
@@ -193,7 +193,7 @@ withEnvConfigExt
     -- this will be run in a Docker container.
     -> Maybe (RIO BuildConfig ())
     -- ^ Action to perform after the build.  This will be run on the host
-    -- OS even if Docker is enabled for builds.  The build config is not
+    -- OS even if Docker is enabled for builds.  The env config is not
     -- available in this action, since that would require build tools to be
     -- installed on the host OS.
     -> RIO BuildConfig a