+++ src/main/Main.hs
@@ -70,7 +70,6 @@ import           Stack.Ghci
 import           Stack.Hoogle
 import           Stack.Ls
 import qualified Stack.IDE as IDE
-import qualified Stack.Image as Image
 import           Stack.Init
 import           Stack.New
 import           Stack.Options.BuildParser
@@ -467,23 +466,6 @@ commandLineHandler currentDir progName isInterpreter = complicatedOptions
                         "Sets a field in the project's stack.yaml to value"
                         cfgSetCmd
                         configCmdSetParser)
-        addSubCommands'
-            Image.imgCmdName
-            "Subcommands specific to imaging"
-            (addCommand'
-                 Image.imgDockerCmdName
-                 "Build a Docker image for the project"
-                 imgDockerCmd
-                 ((,) <$>
-                  boolFlags
-                      True
-                      "build"
-                      "building the project before creating the container"
-                      idm <*>
-                  many
-                      (textOption
-                           (long "image" <>
-                            help "A specific container image name to build"))))
         addSubCommands'
           "hpc"
           "Subcommands specific to Haskell Program Coverage"
@@ -973,18 +955,6 @@ dockerCleanupCmd cleanupOpts =
 
 cfgSetCmd :: ConfigCmd.ConfigCmdSet -> RIO Runner ()
 cfgSetCmd = withGlobalProject . withConfig . cfgCmdSet
-
-imgDockerCmd :: (Bool, [Text]) -> RIO Runner ()
-imgDockerCmd (rebuild,images) = withConfig $ withBuildConfig $ do
-    mProjectRoot <- view $ configL.to configProjectRoot
-    withEnvConfigExt
-        NeedTargets
-        defaultBuildOptsCLI
-        Nothing
-        (\lk ->
-              do when rebuild $ Stack.Build.build Nothing lk
-                 Image.stageContainerImageArtifacts mProjectRoot images)
-        (Just $ Image.createContainerImageFromStage mProjectRoot images)
 
 -- | Project initialization
 initCmd :: InitOpts -> RIO Runner ()